Add WP-hosted font URLs to the template's main.css file (#24)
## What does this PR do?
I was hosting the fonts in the GitHub repository and trying to reference
those `otf` files in the previously available `fonts.css` file. This did
not work because GitHub is not a place that hosts files for development,
so I moved the `.otf` files to the WP media folder and now reference
their static URLs within the template's `main.css` folder. We should be
all good regarding baseline standards now.

## How was the functionality tested and verified?
- [x] I verified that the imports were being loaded on a test visual via
the `Network` tab in the developer window of Chrome.