Adds vote deductions to Election Results (#42)
## What does this PR do?
Adds a 3% vote deduction to two candidates, Demetrius Daniel and Trinity
Bledsoe. This is a strange rule so the simplest way to visualize that I
landed on was to deduct 3% of a candidate's total votes (I rounded to
the nearest whole number. This doesn't affect any rankings) and subtract
that from their ballot votes. I then added an asterisk to their name in
the chart which is explained in a caption. Logic is also added as
comments in code.

## How was the functionality tested and verified?
- [x] The visual was approved by the story's stakeholders.

<img width="1503" alt="Screenshot 2023-10-30 at 11 14 15 AM"
src="https://github.com/chicagomaroon/data-visualizations/assets/45241656/fee9642b-90a5-4172-bf54-cbc98d083201">